Fullpos halo management : ======================= - Revisit the halos management of Aladin-Fullpos in order to allow interpolations over C+I+E. This modification allows a change of resolution of Aladin without reducing the domain area. - Recode LWIDER_DOM. The gridpoint outside the model domain are virtually moved to the limits of the model domain. This modification can have a small impact on the results because the model domain is actually slightly larger than it was thought to be in the previous cycles.
Interoperability aspects :
========================
- Allow NFPCLI=2 again :
NFPCLI=2 is equivalent to NFPCLI=1 but it allows to read more fields that just orography and land-sea mask.
NFPCLI=1 is reserved ionly for orography and land-sea mask in output.
Beware that with this modset certain namelists might need to be changed from NFPCLI=1 to NFPCLI=2.
- Extend VCLIX with more fields
- Flexible FFT package (FFT992 or FFTW) per domain of Fullpos (namelist variable NFPFFTW=0 or 1)
- Use "ICMUA${CNMEXP}INIT" instead of "ICMGG${CNMEXP}INIUA" for gridpoint upper air GRIB2 fields in input of the configuration 903. This modification enables to run the configuration 903 in server mode together with LECMWF=.TRUE.
- Remove LGRBOP and replace by LARPEGEF
- miscellaneous other enhancements for interoperability GRIB2 vs FA
- namelist key LLMODGRIN in conf. 901 to modify fields in input like what is done in IFS.
- printout norms of Fullpos auxilary fields if NFPSURFEX==1
- add PRVRSMIN in surf_inq
- use option 1 with the proper arguments instead of 7 for .not.LCRITSNOWTEMP

Fullpos refactoring :
===================
- Pass the output date in argument.
- Move suoph up from su0yomb to su0yoma and move down sufa
- Move NFPSLWIDE to NAMFPSC2
- Remove suarpio

Bugfixes :
========
Bugfix for LIOLEVG. LIOLEVG can be used to save memory and CPU time in the task "couplingsurf".
Bugfix for the case NFPBW* > 0 with the biperiodicization of Boyd
Bugfix for the configuration 901 when a parameter was unattended or had an incorrect resolution
Bugfix on the call to Fullpos monitoring file
Bugfix to support fullpos server in the task couplingsurf
Bugfix for Surfex compilation
Bugfix for the support of gri_api 2 in FA.
Portability fix for Cray

Optimizations :
=============
- Optimization of the biperiodicization of Boyd
- Optimization of RTTOV interace to Fullpos.
- NJPPF in namelist NAMSATS. Default is 50 (same order of magnitude than NPROMA) to limit
 allocation/deallocation cycles, except for NECSX (256) and Cray (8, as before).
- Optimization by directive "!DEC$ OPTIMIZE:3" for certain subroutines which do take advantage of it.
Note that is directive is not taken into account if the general optimization level is less or equal to 2.
It needs a specific compilation script.
- More use of memcpy
- More use or mpi_waitany
- More use of automatic arrays + overlap send/recv with pack/unpack
- Optimization of RRTM
- Reduce FFTs overhead

Compile time optimization :
=========================
- Split HPOS
- reduce optimization level in setup of ECMWF physics