Add ability to call ecTrans with two "call modes"
INV_TRANS and DIR_TRANS support two ways of passing arrays:
1) PSPVOR, PSPDIV, PSPSCALAR <=> PGP
2) PSPVOR, PSPDIV <=> PGPUV; PSPSC3A <=> PGP3A; PSPSC2 <=> PGP2

Previously we only supported the second one. With this commit you can
now also call the first style. In the IFS we use both cases in different
places so it's important to have both cases covered.

We can now extend the test suite so more of the code has coverage.

In this commit I also tidied up a few things:
- All spectral and grid point work arrays are separately allocated,
  rather than relying on big work arrays to store everything and
  pointers to access slices.
- I deleted a couple unused variables.