Always call GSTATS in GSTATS_NVTX (#230)
In certain subroutines we overload `GSTATS` with `GSTATS_NVTX` which is intended to collect caliper statistics for Nvidia's NVTX library then call `GSTATS` as usual. The latter was inside an `#if defined(__NVCOMPILER)` which means `GSTATS` was not called for other platforms as it should be. I've moved it outside so `GSTATS` is called as usual for other platforms, like for CCE on LUMI-G.