Fix argument checks (#222)
* Improve DIR_TRANS consistency checks

* Improve INV_TRANS consistency checks

* Improve GATH_GRID consistency checks

* Provide more accurate error message in case of wrong KVSET values

* Undo consistency check in INV_TRANS

This check was failing when "--vordiv --uvders" was enabled. There may
be a case in the IFS when we pass a PGPUV array with more wind elements
than required to INV_TRANS. This is certainly the case in the benchmark
program, although I think that's just because the code is wrong. The
--vordiv option doesn't enable grid point vorticity and divergence, as
one would expect from the way the PGPUV array is sized. It just enables
transforming of vorticity and divergence to U and V.

* Improve DIST_GRID consistency checks