Make gpdata truly reduced gridded
Previously we were passing a fake fully gridded input data array which
had zeros in grid points that weren't present in the reduced Gaussian
grid and we relied on ecTrans' loose array dimension checking.

Now that we strictly enforce all array dimensions to be consistent with
NPROMA, NGPTOT etc. we must pass an array that is the proper size for a
reduced Gaussian grid.