Revert optimisation of DIST_GRID
This is a squash of two reverts:

Revert "indentation of OMP directives"

This reverts commit 2fc75f6ec5caf2f6eef79f6b36a9bdcf3dbeb5c6.

Revert "re-phase an optimisation lost after cycle 48"

This reverts commit 861f5e7fb942b3af8012a01cd3bf4e7889816c29.

These two commits were introduced with PR#226
(https://github.com/ecmwf-ifs/ectrans/pull/226).

Unfortunately this PR broke bit-reproducibility in IFS experiments under
certain conditions. The tests suggested a potential bug, either in how
DIST_GRID is called in the IFS or within the proposed optimisations to
DIST_GRID. Until this is resolved, we must remove these commits to
ensure a stable release of ecTrans.

A future release will hopefully re-include the DIST_GRID optimisations
in PR#226 with a bugfix or workaround.