Allow UV derivatives even when grid point vor/div is disabled
The meaning of --vordiv changed in PR #122. Previously --vordiv meant
"include vorticity and divergence in the spectral transform". I.e., pass
vor/div to INV_TRANS, and pass their equivalents, U/V to DIR_TRANS. I
didn't think this was a useful option, as we basically always want to
include vorticity and divergence. Instead I changed it to mean "make
INV_TRANS compute grid point vorticity and divergence, as well as U/V".
I think this is more useful.

Previously we had to disable UV derivatives if the --vordiv option
wasn't passed, because of course it wouldn't be possible to compute U/V
derivatives without an input spectral vor/div. With the new meaning of
--vordiv, U/V derivatives can always be computed (if --uvders is passed)
because vor/div is always passed to INV_TRANS.

Maybe we should change the name of --vordiv though. Something like
--vordivgp.