Revert "Fix for cublas link failures (#302)"
This reverts commit d93d764017c95e8f252aa2925a6a900125be6943.

The reverted fix is a possible workaround to a compiler bug when having the `-cuda` flag,
which adds implicitly `-lcublas` and `-lcurand` on the link line without the link
directories. This is the case in ectrans version 1.6.2.
However already in the develop branch, with PR #261, we already removed this `-cuda` flag.

A possible fix for 1.6.2 and any package using `-cuda` could instead be handled in a toolchain file e.g. by using
CMAKE_Fortran_IMPLICIT_LINK_DIRECTORIES.