Remove MPI condition for GPU feature
In fact you don't need to build against MPI to run the GPU version
with a single MPI task. If ecTrans is built against FIAT with the MPI
feature disabled, ecTrans now inherits this. If running with a single
task, MPI calls are simply bypassed following the logic of the
communication routines. If running with more than one task, and abort is
thrown.