Tweak CMake features (#330)
* Make MPI default ON

* Move CPU feature up and make it condition for MKL feature

* Make sure one of the two precision features is enabled

* Make MPI condition for GPU feature

Currently we call raw MPI routines in the GPU version, so this is a
required feature.

* Make GPU a condition for GPU_STATIC feature

* Remove tabs

* Remove independent ecTrans MPI feature

* Remove unnecessary MPI CXX component for GPU_AWARE_MPI

* Fix CUDA reference

* Remove MPI condition for GPU feature

In fact you don't need to build against MPI to run the GPU version
with a single MPI task. If ecTrans is built against FIAT with the MPI
feature disabled, ecTrans now inherits this. If running with a single
task, MPI calls are simply bypassed following the logic of the
communication routines. If running with more than one task, and abort is
thrown.

* Revert change to MPI linking of GPU version

Actually MPI is still required for running with more than one task even
when GPU-aware communication is disabled.