Move INIGPTR to common
This is precision agnostic so pretty straightforward to move to common.
I just had to modify the signature so that NPROMA and NGPBLKS are passed
as arguments. This is necessary because TPM_TRANS is not in common, so
we cannot simply take these variables from that module.

Note that the GPU version could have been deleted in any case as it
isn't used at all.